adding readme, changes to UI and app. Trying to make code look better
